[SR-152] update address controller

+++ smart-rent/src/main/java/com/smartrent/dto/response/ListingResponseWithAdmin.java
@@ -68,7 +68,7 @@ public class ListingResponseWithAdmin {
 
     String priceUnit;
 
-    Integer addressId;
+    Long addressId;
 
     Float area;
